[QOL-7970] split off unit testing for Pylons plugin
- fake an older CKAN version so we can test Pylons
